% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{260ALS_genes}
\alias{260ALS_genes}
\title{Example data for Amyotrophic lateral sclerosis gene set}
\format{
a vector that contains 260 genes related to Amyotrophic lateral sclerosis
}
\usage{
260ALS_genes
}
\description{
Example data for Amyotrophic lateral sclerosis gene set
}
\keyword{datasets}
