% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_marker_genes_of_mouse_brain}
\alias{cell_type_marker_genes_of_mouse_brain}
\title{Example data for marker gene lists of two cell type in mouse brain}
\format{
a dataframe contains statistic information (p value and adjusted p value) and marker gene namse of two cell type (L6 IT CTX and Oligodendrocyte)
}
\usage{
cell_type_marker_genes
}
\description{
Example data for marker gene lists of two cell type in mouse brain
}
\keyword{datasets}
