% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustered_mean_Exp.R
\name{Clustered_mean_Exp}
\alias{Clustered_mean_Exp}
\title{Calculating the mean expression of all genes in each clusters}
\usage{
Clustered_mean_Exp(clusteredExp)
}
\arguments{
\item{clusteredExp}{a list containing expression matrixes divided by clusters}
}
\value{
A dataframe containing mean expression of all genes in each clusters
}
\description{
Calculate the mean expression of each genes in each clusers separately
}
\examples{
clustered_exp <- getClusteredExp(mouse_brain,exp)
clustered_mean_exp <- Clustered_mean_Exp(clustered_exp)
}
