% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculate information entropies of all genes}
\usage{
entropy(percentage)
}
\arguments{
\item{percentage}{the percentage dataframe which to use}
}
\value{
A dataframe containing information entropies of all genes
}
\description{
calculate information entropies of all genes according to the degree of dispersion
}
\examples{
entropy_gene <- entropy(percentage_gene)
}
