% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genefilter.R
\name{genefilter}
\alias{genefilter}
\title{filtering out genes which expression levels are too low}
\usage{
genefilter(clusteredMeanExp, threshhold = 0.2)
}
\arguments{
\item{clusteredMeanExp}{a dataframe containing mean expression of all genes in each clusters, rows are genes, columns are clusters}

\item{threshhold}{Limits for filtering genes.Default is 0.2}
}
\value{
A dataframe containing mean expression of genes which expression levels are hingher than threshhold
}
\description{
Filter out genes which mean expression levels are lower than threshhold
}
\examples{
clustered_mean_exp_filtered <- genefilter(clustered_mean_exp, threshhold = 0.2)
}
