% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusteredExp.R
\name{getClusteredExp}
\alias{getClusteredExp}
\title{Getting expression data grouped by clusters}
\usage{
getClusteredExp(SeuratObject, ExpMatrix)
}
\arguments{
\item{SeuratObject}{a Seurat object of spatial transcriptomics}

\item{ExpMatrix}{a matrix containing the expression values of all genes in all spots}
}
\value{
A list containing expression matrixes divided by clusters
}
\description{
This function divides the expression matrix into different matrixes according to clusters
}
\examples{
exp <- getExpMatrix(mouse_brain)
clustered_exp <- getClusteredExp(mouse_brain,exp)
}
