% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentOutput.R
\name{getEnrichmentOutput}
\alias{getEnrichmentOutput}
\title{Getting the enrichment output of target gene set}
\usage{
getEnrichmentOutput(target_gama, random_gama)
}
\arguments{
\item{target_gama}{the weighted deviation accumulation dataframe of target gene set}

\item{random_gama}{the list containing 10,000 weighted deviation accumulation dataframe of randomly sampling gene set}
}
\value{
A dataframe containing the enrichment output of target gene set and associated statistics (pval, fdr, bonferroni, etc.)
}
\description{
get the enrichment output of target gene set by comparing the weighted deviation accumulation of target gene set with randomly sampling gene set
}
\examples{
enrichment_output <- getEnrichmentOutput(target_gama, random_gama)
}
