% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpMatrix.R
\name{getExpMatrix}
\alias{getExpMatrix}
\title{Getting expression matrix}
\usage{
getExpMatrix(SeuratObject)
}
\arguments{
\item{SeuratObject}{a Seurat object of spatial transcriptomics}
}
\value{
A matrix containing the expression values of all genes in all spots
}
\description{
Get the expression matrix from Seurat object of spatial transcriptomics
}
\examples{
data("mouse_brain")
mouse_brain <- SCTransform(mouse_brain, assay = "Spatial") \%>\%
               RunPCA(assay = "SCT", verbose = FALSE) \%>\%
               FindNeighbors(reduction = "pca", dims = 1:30) \%>\%
               FindClusters(resolution = 0.8) \%>\%
               RunUMAP(reduction = "pca", dims = 1:30)
exp <- getExpMatrix(mouse_brain)
}
