% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGama.R
\name{getGama}
\alias{getGama}
\title{Calculate the weighted deviation accumulation of gene set}
\usage{
getGama(geneset_delta, type = c("target", "random"), cores = n)
}
\arguments{
\item{geneset_delta}{a dataframe containing delta values of genes in gene set, the last column is the weight of gene}

\item{type}{provided gene set type. Available options are "target" for taget gene set; "random" for randomly sampling gene set}

\item{cores}{number of cores used for parallel computation}
}
\value{
A dataframe or a list containing the weighted deviation accumulation of gene set in each clusters
}
\description{
calculate the weighted deviation accumulation of provided gene set in each clusters
}
\examples{
target_gama <- getGama(target_delta, type = "target")
random_gama <- getGama(random_delta, type = "random", cores = 10)
}
