% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomSample.R
\name{getRandomSample}
\alias{getRandomSample}
\title{Getting the randomly sampling gene set and corresponding delta matrix}
\usage{
getRandomSample(delta, target_geneset)
}
\arguments{
\item{delta}{the delta dataframe of all genes}

\item{target_geneset}{the target gene set which to test}
}
\value{
A list containing 10,000 randomly sampling gene set and corresponding delta matrix
}
\description{
get 10,000 randomly sampling gene set with same length of target gene set and corresponding delta matrix
}
\examples{
random_delta <- getRandomSample(delta, target_geneset)
}
