% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_brain}
\alias{mouse_brain}
\title{Example data for a mouse brain}
\format{
a Seurat object of a spatial transcriptom data of a sagittal section of the anterior half of the mouse brain
}
\usage{
mouse_brain
}
\description{
Example data for a mouse brain
}
\keyword{datasets}
