% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{percentage}
\alias{percentage}
\title{Percentage calcultion}
\usage{
percentage(rescale)
}
\arguments{
\item{rescale}{a rescale dataframe which to use}
}
\value{
A dataframe containing expression percentages of all genes in each clusters
}
\description{
calculate the expression percentages of all genes in each clusters
}
\examples{
percentage_gene <- percentage(rescale_gene)
}
