% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale calculation}
\usage{
rescale(delta)
}
\arguments{
\item{delta}{the delta dataframe which to use}
}
\value{
A dataframe containing rescale values
}
\description{
rescale the delta values
}
\examples{
rescale_gene <- rescale(delta)
}
