% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{Calculate the weight of all genes}
\usage{
weight(entropy)
}
\arguments{
\item{entropy}{the entropy dataframe which to use}
}
\value{
A dataframe containing weight of all genes
}
\description{
calculate the weight of all genes according to the degree of dispersion
}
\examples{
weight_gene <- weight(entopy_gene)
}
